+++ MonsterChase/RandomController.h
@@ -7,7 +7,7 @@ public:
 	RandomController();
 	void SetActor(Actor * i_pActor) override { m_Actor = i_pActor; }
 	void UpdateActor() override;
-	Point2D<int> GetRandomMovement();
+	Point2D<float> GetRandomMovement();
 	~RandomController();
 private:
 	Actor * m_Actor;
